+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added default keys for the `getRefreshRateInfo()` that are always returned.
   - `platform`: (string) The platform the app is running on, e.g. "android", "ios", "windows", "linux", "macos"
   - `supported`: (bool) Whether refresh rate control is supported on the current platform
+* Aliased 'currentFramesPerSecond', 'currentRefreshRate' for iOS and Android so both will give you the same value.
 # 0.0.4+1
 
 * Lower Dart SDK minimum version to `3.7.0`